menampilkan data permasalahan di strategic arah kebijakan opd, dan hapus tabel tujuan
